test opening in read only mode
